hint: move the hint package into the ast package
Since hint needs to be imported by ast, and ast is not allowed to
import anything from the main earthly module, the hint package
must be in the ast package or be its own module.
