+++ ast/antlrhandler/error_listener.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 
 	"github.com/antlr/antlr4/runtime/Go/antlr/v4"
+	"github.com/earthly/earthly/ast/hint"
 	"github.com/earthly/earthly/ast/parser"
-	"github.com/earthly/earthly/util/hint"
 	"github.com/pkg/errors"
 )
