+++ ast/hint/error.go
