+++ ast/parse_errors_test.go
@@ -49,9 +49,10 @@ test:
     IF $foo END
 `,
 			expectedHint: `
-Hints:
+  Hints:
   - I couldn't find a pattern that completes the current statement - check your quote pairs, paren pairs, and newlines
-  - I parsed 'END' as a word, but it looks like it should be a keyword - is it on the wrong line?`,
+  - I parsed 'END' as a word, but it looks like it should be a keyword - is it on the wrong line?
+`,
 		},
 		{
 			name: "key-value with missing EQUALS",
@@ -63,8 +64,8 @@ test:
     LABEL a
 `,
 			expectedHint: `
-Hints:
-  - I got lost looking for '=' - did you define a key/value pair without a value?`,
+  Hint: I got lost looking for '=' - did you define a key/value pair without a value?
+`,
 		},
 	} {
 		tt := tt
